Fix ignoring parameters
In all other places the SwaggerIgnoreAttribute is matched exactly by name; but for parameters a type that inherits from SwaggerIgnoreAttribute is expected. This makes hiding parameters tedious as it requires creating a new attribute type and produces inconsistency in the way that annotations are handled. This commit makes annotations consistent and allows [SwaggerIgnore] on parameters.